gst-utils: streamproducer: add add_consumer_with()
A consumer can be added to a `StreamProducer` using `add_consumer()`. This
method configures properties on the provided `appsrc`, possibly overriding
values defined by the caller.

This commit adds a new `add_consumer_with()` method which
initialises the `appsrc` in a state suitable for its use as a consumer,
just like `add_consumer()`, but also applies settings defined in the provided
`ConsumerSettings`. These settings are limited to properties of the internal
queue for now.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer-rs/-/merge_requests/1794>